+++ CHANGES.md
@@ -12,6 +12,7 @@ For a full list of changes, see the [git commit log][log]
   * Enable service recovery if service terminates without stop signal
   * Do not wait for shutdown process cleanup on timeout
   * Kill all child processes if service terminates without stop signal
+  * Drop useless -c command option
 
 ### New Features